Added more info about the repo and files
More in depth about the files in the repo, the data that is acquired with the python scripts, as well as formatting of that data. Information regarding the Environment Canada Licence was also added along with a link for where documentation is located on the Environment Canada server regarding this data.