Update SWOB-ML Reader.py
Re-integrated the CSV writer, and brought the formatting into line with what you would receive from Environment Canada if you were purchasing data from them. There is still a text file written but it is not as clear as it used to be. I may just remove it in the future or bring the text file back to the old format this this script is just for one XML file at a time anyway.